Converters now take batched norm data, to prevent out of memory. Cleaned up evaluate_batched code (no longer requires knowledge of dataset length). Example projects are cleaned up, and now use batched datasets in conversion and evaluation.
